Renamed TaskInput to TheInput.
